+++ gem/lib/pagy/toolbox/paginators/countless.rb
@@ -6,16 +6,19 @@ class Pagy
 
     # Return the Offset::Countless instance and records
     def paginate(collection, options)
-      if options[:page].nil?
-        page = options[:request].resolve_page(force_integer: false) # accept nil and strings
-        if page.is_a?(String)
-          p, l = page.split(/ /, 2).map(&:to_i) # decoded '+' added by the compose_page_url
-          options[:page] = p if p.positive?
-          options[:last] = l if l&.positive?
+      options[:page] ||= options[:request].resolve_page(force_integer: false) # accept nil and strings
+      if options[:page].is_a?(String)
+        page, last     = options[:page].split(/ /, 2).map(&:to_i) # decoded '+' added by the compose_page_url
+        options[:page] = page
+        if last&.positive?
+          options[:last] = last
+        else
+          # Legacy, last-less page links are handled by starting from the first page
+          options[:page] = nil
         end
       end
       options[:limit] = options[:request].resolve_limit
-      pagy = Offset::Countless.new(**options)
+      pagy            = Offset::Countless.new(**options)
       [pagy, pagy.records(collection)]
     end
   end